# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Pixelate",
        "description": "Documentation for Pixelate app",
        "version": "1.0.2"
    },
    "servers": [
        {
            "url": "http://localhost",
            "description": "Server"
        }
    ],
    "paths": {
        "/api/posts": {
            "get": {
                "tags": [
                    "Post"
                ],
                "summary": "Get all Posts",
                "description": "Get all Posts available online (set to live )",
                "operationId": "App\\Http\\Controllers\\API\\PostAPIController::index",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Post"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Display a listing of the Post.\nGET|HEAD /posts",
                                            "type": "string",
                                            "example": "Not Found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/posts/{user_id}": {
            "post": {
                "tags": [
                    "Post"
                ],
                "summary": "Create Post",
                "description": "Create A Post",
                "operationId": "App\\Http\\Controllers\\API\\PostAPIController::store",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "number",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "description": "Pass Post data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "caption": {
                                        "description": "Store a newly created Post in storage.\nPOST /posts",
                                        "type": "string",
                                        "example": "Here"
                                    },
                                    "imageUrl": {
                                        "description": "Store a newly created Post in storage.\nPOST /posts",
                                        "type": "string",
                                        "example": "some image url"
                                    },
                                    "location": {
                                        "description": "Store a newly created Post in storage.\nPOST /posts",
                                        "type": "string",
                                        "example": "Dusseldorf"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "Store a newly created Post in storage.\nPOST /posts",
                                            "type": "string",
                                            "example": "false"
                                        },
                                        "message": {
                                            "description": "Store a newly created Post in storage.\nPOST /posts",
                                            "type": "string",
                                            "example": "Post not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/posts/{id}": {
            "get": {
                "tags": [
                    "Post"
                ],
                "summary": "Get Post by Id",
                "description": "Get Single Post",
                "operationId": "App\\Http\\Controllers\\API\\PostAPIController::show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "number",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Post"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "Display the specified Post.\nGET|HEAD /posts/{id}",
                                            "type": "string",
                                            "example": "false"
                                        },
                                        "message": {
                                            "description": "Display the specified Post.\nGET|HEAD /posts/{id}",
                                            "type": "string",
                                            "example": "Post not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Post"
                ],
                "summary": "Update Post",
                "description": "Update A Post by Id",
                "operationId": "App\\Http\\Controllers\\API\\PostAPIController::update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "number",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "description": "Pass Post data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "caption": {
                                        "description": "Update the specified Post in storage.\nPUT/PATCH /posts/{id}",
                                        "type": "string",
                                        "example": "Here"
                                    },
                                    "location": {
                                        "description": "Update the specified Post in storage.\nPUT/PATCH /posts/{id}",
                                        "type": "string",
                                        "example": "Dusseldorf"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "Update the specified Post in storage.\nPUT/PATCH /posts/{id}",
                                            "type": "string",
                                            "example": "true"
                                        },
                                        "message": {
                                            "description": "Update the specified Post in storage.\nPUT/PATCH /posts/{id}",
                                            "type": "string",
                                            "example": "Post deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "Update the specified Post in storage.\nPUT/PATCH /posts/{id}",
                                            "type": "string",
                                            "example": "false"
                                        },
                                        "message": {
                                            "description": "Update the specified Post in storage.\nPUT/PATCH /posts/{id}",
                                            "type": "string",
                                            "example": "Post not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Post"
                ],
                "summary": "Delete Post",
                "description": "Delete A Post by Id",
                "operationId": "App\\Http\\Controllers\\API\\PostAPIController::destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "number",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "Remove the specified Post from storage.\nDELETE /posts/{id}",
                                            "type": "string",
                                            "example": "true"
                                        },
                                        "message": {
                                            "description": "Remove the specified Post from storage.\nDELETE /posts/{id}",
                                            "type": "string",
                                            "example": "Post deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "Remove the specified Post from storage.\nDELETE /posts/{id}",
                                            "type": "string",
                                            "example": "false"
                                        },
                                        "message": {
                                            "description": "Remove the specified Post from storage.\nDELETE /posts/{id}",
                                            "type": "string",
                                            "example": "Post not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Comment": {
                "properties": {
                    "id": {
                        "description": "App\\Models\\Comment",
                        "type": "integer",
                        "readOnly": "true",
                        "example": "1"
                    },
                    "created_dates": {
                        "description": "App\\Models\\Comment",
                        "properties": {
                            "created_at_human": {
                                "description": "Date Created formatted",
                                "example": "52 minutes ago"
                            },
                            "created_at": {
                                "description": "Raw unfarmatted Date ",
                                "example": "2020-11-09T20:04:11.000000Z"
                            }
                        },
                        "type": "object"
                    },
                    "updated_dates": {
                        "description": "App\\Models\\Comment",
                        "properties": {
                            "updated_at_human": {
                                "description": "Date Updated formatted",
                                "example": "52 minutes ago"
                            },
                            "updated_at": {
                                "description": "Raw unfarmatted update date ",
                                "example": "2020-11-09T20:04:11.000000Z"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Comment"
                }
            },
...
            "Post": {
                "properties": {
                    "id": {
                        "description": "Class Post",
                        "type": "integer",
                        "readOnly": "true",
                        "example": "1"
                    },
                    "caption": {
                        "description": "Class Post",
                        "type": "string",
                        "example": "the sea life"
                    },
                    "location": {
                        "description": "Class Post",
                        "type": "string",
                        "example": "Dusseldorf"
                    },
                    "imageUrl": {
                        "description": "Class Post",
                        "type": "string",
                        "example": "http://localhost:8000/storage/3B5wcGnEbnFaAYsjMxk5P1V1fAGKaVhviC3EO0Gd.png"
                    },
                    "created_dates": {
                        "description": "Class Post",
                        "properties": {
                            "created_at_human": {
                                "description": "Date Created formatted",
                                "example": "52 minutes ago"
                            },
                            "created_at": {
                                "description": "Raw unfarmatted Date ",
                                "example": "2020-11-09T20:04:11.000000Z"
                            }
                        },
                        "type": "object"
                    },
                    "updated_dates": {
                        "description": "Class Post",
                        "properties": {
                            "updated_at_human": {
                                "description": "Date Updated formatted",
                                "example": "52 minutes ago"
                            },
                            "updated_at": {
                                "description": "Raw unfarmatted update date ",
                                "example": "2020-11-09T20:04:11.000000Z"
                            }
                        },
                        "type": "object"
                    },
                    "user": {
                        "$ref": "#/components/schemas/User"
                    },
                    "comment": {
                        "$ref": "#/components/schemas/Comment"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Post"
                }
            },
            "User": {
                "required": [
                    "password"
                ],
                "properties": {
                    "id": {
                        "description": "App\\Models\\User",
                        "type": "integer",
                        "readOnly": "true",
                        "example": 1
                    },
                    "name": {
                        "description": "App\\Models\\User",
                        "type": "string",
                        "maxLength": 32,
                        "example": "Mittie Morissette"
                    },
                    "username": {
                        "description": "App\\Models\\User",
                        "type": "string",
                        "maxLength": 32,
                        "example": "pierce"
                    },
                    "email": {
                        "description": "User unique email address",
                        "type": "string",
                        "format": "email",
                        "readOnly": "true",
                        "example": "fanny256@email.com"
                    },
                    "email_verified_at": {
                        "description": "Datetime marker of verification status",
                        "type": "string",
                        "format": "date-time",
                        "readOnly": "true",
                        "example": "2019-02-25 12:59:20"
                    },
                    "two_factor_secret": {
                        "description": "App\\Models\\User",
                        "type": "number",
                        "example": null
                    },
                    "two_factor_recovery_codes": {
                        "description": "App\\Models\\User",
                        "type": "number",
                        "example": null
                    },
                    "tagline:": {
                        "description": "App\\Models\\User",
                        "type": "string",
                        "maxLength": 32,
                        "example": "Producer:"
                    },
                    "location": {
                        "description": "App\\Models\\User",
                        "properties": {
                            "type": {
                                "description": "App\\Models\\User",
                                "type": "string",
                                "example": "point"
                            },
                            "coordinates": {
                                "description": "App\\Models\\User",
                                "example": [
                                    8.503972,
                                    51.017243
                                ]
                            }
                        },
                        "type": "object"
                    },
                    "formatted_address": {
                        "description": "App\\Models\\User",
                        "example": "811 Sibyl Bypass Suite 783\\n New Rita, AL 48220-0930"
                    },
                    "available_to_hire": {
                        "description": "App\\Models\\User",
                        "type": "boolean",
                        "example": 1
                    },
                    "about": {
                        "description": "App\\Models\\User",
                        "type": "string",
                        "example": "VERY deeply with a soldier on each."
                    },
                    "trade_id": {
                        "description": "App\\Models\\User",
                        "type": "number",
                        "example": null
                    },
                    "current_team_id": {
                        "description": "App\\Models\\User",
                        "type": "number",
                        "example": null
                    },
                    "profile_photo_path": {
                        "description": "App\\Models\\User",
                        "type": "number",
                        "example": null
                    },
                    "created_at": {
                        "$ref": "#/components/schemas/BaseModel/properties/created_at"
                    },
                    "updated_at": {
                        "$ref": "#/components/schemas/BaseModel/properties/updated_at"
                    },
                    "photo_url:": {
                        "description": "App\\Models\\User",
                        "type": "string",
                        "maxLength": 32,
                        "example": "https://www.gravatar.com/avatar/97bd1823e00f02eb71a0b709425152a7jpg?s=200&d=mm:"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "User"
                }
            }
        },
        "securitySchemes": {
            "token": {
                "type": "apiKey",
                "description": "Example of Auth Header Token (Copy and paste below to authorize)=>\nBearer eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpc3MiOiJodHRwOlwvXC9sb2NhbGhvc3Q6ODAwMFwvYXBpXC9sb2dpbiIsImlhdCI6MTYwNTM0ODU1OSwiZXhwIjoxNjA1MzUyMTU5LCJuYmYiOjE2MDUzNDg1NTksImp0aSI6IjlqMDN0MVVISThoNjlhRUsiLCJzdWIiOjEsInBydiI6IjIzYmQ1Yzg5NDlmNjAwYWRiMzllNzAxYzQwMDg3MmRiN2E1OTc2ZjcifQ.dn45EdZMatpW2zAQcZklDS2902mwVNx44TbkDiiGLXY",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "security": [
        []
    ]
}